feat(772): Sweep tab — native lineage graph (vis-network) + shared detail component
A Board | Graph toggle on the Fault Sweep tab. The graph renders the sweep's
lineage DAG: nodes = experiments (colored by verdict → status), edges = parent
(seed → isolation fan → bisect, arrows) + group (A/B, dashed); hierarchical
top-down layout, fit-to-view. Click a node → its full detail in the side panel.

- useChartJs.ts: ensureVisNetwork() — CDN lazy-load of vis-network@9.1.9
  (mirrors the existing vis-timeline loader; not bundled).
- SweepJobDetail.vue (new): the per-job detail block (recipe/provenance/outcome/
  next-action) extracted so the board card AND the graph node panel share it
  (single source of truth, no duplication).
- Sweep.vue: viewMode toggle, buildGraph()/nodeColor(), board uses the component.

Verified live on test-dev: vis-network loads from CDN, the graph renders the
seed-pyramid → isolation-fan/bisect tree with verdict colors, and the legend +
click-to-detail side panel work.
